requirements.txt
Django==1.8.2
Pillow==2.8.2
PyJWT==1.3.0
PyYAML==3.11
Werkzeug==0.10.4
amqp==1.4.6
anyjson==0.3.3
argparse==1.1
arrow==0.5.4
billiard==3.3.0.20
celery==3.1.18
crcmod==1.7
cssselect==0.9.1
django-activity-stream==0.5.1
django-annoying==0.8.2
django-bower==5.0.4
django-celery==3.1.16
django-cors-headers==1.1.0
django-environ==0.3.0
django-extensions==1.5.5
django-filter==0.10.0
django-inmemorystorage==0.1.1
django-jsonfield==0.9.13
django-pipeline==1.5.1
django-rest-swagger==0.3.2
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
drf-extensions==0.2.7
easy-thumbnails==2.2
freezegun==0.3.3
future==0.14.3
futures==3.0.2
icalendar==3.9.0
kombu==3.0.26
lxml==3.4.4
mock==1.0.1
mock-django==0.6.8
model-mommy==1.2.4
oauthlib==0.7.2
pyquery==1.2.9
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.4
requests==2.7.0
requests-mock==0.6.0
requests-oauthlib==0.5.0
rollbar==0.9.11
six==1.9.0
wsgiref==0.1.2
